# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/api/students/": {
      "get": {
        "tags": [
          "students"
        ],
        "summary": "Read Students",
        "operationId": "read_students_api_students__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StudentResponse"
                  },
                  "type": "array",
                  "title": "Response Read Students Api Students  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "students"
        ],
        "summary": "Create Student",
        "operationId": "create_student_api_students__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/students/{student_id}": {
      "get": {
        "tags": [
          "students"
        ],
        "summary": "Read Student",
        "operationId": "read_student_api_students__student_id__get",
        "parameters": [
          {
            "name": "student_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Student Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "students"
        ],
        "summary": "Update Student",
        "operationId": "update_student_api_students__student_id__put",
        "parameters": [
          {
            "name": "student_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Student Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "students"
        ],
        "summary": "Delete Student",
        "operationId": "delete_student_api_students__student_id__delete",
        "parameters": [
          {
            "name": "student_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Student Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/rooms/": {
      "get": {
        "tags": [
          "rooms"
        ],
        "summary": "Read Rooms",
        "operationId": "read_rooms_api_rooms__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RoomResponse"
                  },
                  "type": "array",
                  "title": "Response Read Rooms Api Rooms  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "rooms"
        ],
        "summary": "Create Room",
        "operationId": "create_room_api_rooms__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rooms/{room_id}": {
      "get": {
        "tags": [
          "rooms"
        ],
        "summary": "Read Room",
        "operationId": "read_room_api_rooms__room_id__get",
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Room Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "rooms"
        ],
        "summary": "Update Room",
        "operationId": "update_room_api_rooms__room_id__put",
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Room Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "rooms"
        ],
        "summary": "Delete Room",
        "operationId": "delete_room_api_rooms__room_id__delete",
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Room Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rooms/by_floor/{floor_id}": {
      "get": {
        "tags": [
          "rooms"
        ],
        "summary": "Read Rooms By Floor",
        "operationId": "read_rooms_by_floor_api_rooms_by_floor__floor_id__get",
        "parameters": [
          {
            "name": "floor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Floor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoomResponse"
                  },
                  "title": "Response Read Rooms By Floor Api Rooms By Floor  Floor Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/violations/": {
      "get": {
        "tags": [
          "violations"
        ],
        "summary": "Read Violations",
        "operationId": "read_violations_api_violations__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ViolationResponse"
                  },
                  "type": "array",
                  "title": "Response Read Violations Api Violations  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "violations"
        ],
        "summary": "Create Violation",
        "operationId": "create_violation_api_violations__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ViolationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViolationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/violations/{violation_id}": {
      "get": {
        "tags": [
          "violations"
        ],
        "summary": "Read Violation",
        "operationId": "read_violation_api_violations__violation_id__get",
        "parameters": [
          {
            "name": "violation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Violation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViolationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "violations"
        ],
        "summary": "Update Violation",
        "operationId": "update_violation_api_violations__violation_id__put",
        "parameters": [
          {
            "name": "violation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Violation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ViolationCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViolationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "violations"
        ],
        "summary": "Delete Violation",
        "operationId": "delete_violation_api_violations__violation_id__delete",
        "parameters": [
          {
            "name": "violation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Violation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViolationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "RoomCreate": {
        "properties": {
          "floor_id": {
            "type": "integer",
            "title": "Floor Id"
          },
          "room_type": {
            "type": "string",
            "title": "Room Type"
          },
          "room_number": {
            "type": "integer",
            "title": "Room Number"
          },
          "bed_count": {
            "type": "integer",
            "title": "Bed Count"
          }
        },
        "type": "object",
        "required": [
          "floor_id",
          "room_type",
          "room_number",
          "bed_count"
        ],
        "title": "RoomCreate"
      },
      "RoomResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "floor_id": {
            "type": "integer",
            "title": "Floor Id"
          },
          "room_type": {
            "type": "string",
            "title": "Room Type"
          },
          "room_number": {
            "type": "integer",
            "title": "Room Number"
          },
          "bed_count": {
            "type": "integer",
            "title": "Bed Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "floor_id",
          "room_type",
          "room_number",
          "bed_count"
        ],
        "title": "RoomResponse"
      },
      "StudentCreate": {
        "properties": {
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "title": "Last Name"
          },
          "birth_date": {
            "type": "string",
            "format": "date",
            "title": "Birth Date"
          },
          "contact_info": {
            "type": "string",
            "title": "Contact Info"
          },
          "course": {
            "type": "integer",
            "title": "Course"
          },
          "grup": {
            "type": "integer",
            "title": "Grup"
          },
          "is_non_local": {
            "type": "boolean",
            "title": "Is Non Local"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "gender": {
            "type": "string",
            "title": "Gender"
          }
        },
        "type": "object",
        "required": [
          "first_name",
          "last_name",
          "birth_date",
          "contact_info",
          "course",
          "grup",
          "is_non_local",
          "password",
          "gender"
        ],
        "title": "StudentCreate"
      },
      "StudentResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "title": "Last Name"
          },
          "birth_date": {
            "type": "string",
            "format": "date",
            "title": "Birth Date"
          },
          "contact_info": {
            "type": "string",
            "title": "Contact Info"
          },
          "course": {
            "type": "integer",
            "title": "Course"
          },
          "grup": {
            "type": "integer",
            "title": "Grup"
          },
          "is_non_local": {
            "type": "boolean",
            "title": "Is Non Local"
          },
          "gender": {
            "type": "string",
            "title": "Gender"
          }
        },
        "type": "object",
        "required": [
          "id",
          "first_name",
          "last_name",
          "birth_date",
          "contact_info",
          "course",
          "grup",
          "is_non_local",
          "gender"
        ],
        "title": "StudentResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ViolationCreate": {
        "properties": {
          "student_id": {
            "type": "integer",
            "title": "Student Id"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          }
        },
        "type": "object",
        "required": [
          "student_id",
          "description",
          "date"
        ],
        "title": "ViolationCreate"
      },
      "ViolationResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "student_id": {
            "type": "integer",
            "title": "Student Id"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "violation_date": {
            "type": "string",
            "format": "date",
            "title": "Violation Date"
          }
        },
        "type": "object",
        "required": [
          "id",
          "student_id",
          "description",
          "violation_date"
        ],
        "title": "ViolationResponse"
      }
    }
  }
}